% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_fit.R
\name{print.btfit}
\alias{print.btfit}
\title{Print method for btfit objects}
\usage{
\method{print}{btfit}(x, ...)
}
\arguments{
\item{x}{A \code{btfit} object.}

\item{...}{Further arguments (ignored).}
}
\description{
Print method for btfit objects
}
