% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_summary.R
\name{print.summary.btfit}
\alias{print.summary.btfit}
\title{Print method for summary.btfit objects}
\usage{
\method{print}{summary.btfit}(x, digits = 4, ...)
}
\arguments{
\item{x}{A \code{summary.btfit} object.}

\item{digits}{Integer. Number of decimal places to display. Default \code{4}.}

\item{...}{Further arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for summary.btfit objects
}
